feat: split up helpers
